+++ src/commands/methods.js
@@ -1,5 +1,8 @@
 'use strict';
 
-module.exports = api => {
-
+module.exports = {
+    // create
+    ...require('./create/methods'),
+    // deploy
+    ...require('./deploy/methods'),
 };