+++ theme/config/index.js
@@ -1,3 +1,3 @@
 export const BEIAN_LINK = 'http://www.beian.miit.gov.cn/';
-export const POWERBY_LINK = 'https://MicroAppJS.github.io/plugin-vuepress';
+export const POWERBY_LINK = 'https://MicroAppJS.github.io/vuepress';
 export const POWERBY_NAME = 'MicroApp';